Displaying corresponding type of progress
Now the Habit Card displays either a checkbox or a number, depending on the type of the corresponding habit.

When creating two or more enumerable Habits in a row (created one, edited it to give it some value, then created a second one), value from the already created one carries over to the next one.